Extract case_record_flow.rs from turn.rs (parent #680)
Vertical-slice extraction: move the case-record extraction + retrieval
flow out of turn.rs into a new src/agent/loop_run/case_record_flow.rs
sibling module. Same pattern as anti_pattern_flow.

Migrated (all pub(super) free fns):

Entry points:
- maybe_extract_case_record(&mut Agent, &LoopStats, &[String]) -> Option<CaseRecord>
  (Issue #462 post-loop success-condition + persist)
- try_inject_case_retrieval_message(&mut Agent) -> Option<RetrievalInjection>
  (Issue #463 pre-prompt retrieval)

Private helpers:
- finish_case_record_extraction (per-turn cap setter, generic over T)
- persist_case_record (persist + telemetry)

No facade re-export (DR3-001). Caller updates:
- turn.rs: `self.try_inject_case_retrieval_message()` →
  `super::case_record_flow::try_inject_case_retrieval_message(self)`
- actor_loop_flow.rs: `agent.maybe_extract_case_record(&stats, &verify_commands_collected)` →
  `super::case_record_flow::maybe_extract_case_record(agent, &stats, &verify_commands_collected)`

`cargo fix --tests` pruned 2 newly-unused imports from turn.rs's prelude.

turn.rs: 6,884 -> 6,562 (-322 LOC)
Cumulative from 39,489: -32,927 LOC (-83.4%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
